changed all prng's to reference; otherwise it always does the same

+++ free_functions.cpp
@@ -9,7 +9,7 @@ using namespace std;
 double getLogGrowth(const Island& island) noexcept
 {
     return 1.0 - static_cast<double>(island.getNSpecies())
-            / island.getCarryingCap();
+            / static_cast<double>(island.getCarryingCap());
 }
 
 double extractSumOfRates(const Island& island)
